rename score_utils.py -> utils.py
