+++ vllm/entrypoints/pooling/score/protocol.py
@@ -12,7 +12,7 @@ from vllm import PoolingParams
 from vllm.config.pooler import get_use_activation
 from vllm.entrypoints.openai.engine.protocol import OpenAIBaseModel, UsageInfo
 from vllm.entrypoints.pooling.base.protocol import PoolingBasicRequestMixin
-from vllm.entrypoints.pooling.score.score_utils import (
+from vllm.entrypoints.pooling.score.utils import (
     ScoreContentPartParam,
     ScoreMultiModalParam,
 )
